+++ src/main/java/com/example/DeliveryNotificationModule/service/VehicleDeliveryService.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface VehicleDeliveryService {
 
       VehicleDeliveryDto createVehicleDelivery(VehicleDeliveryDto vehicleDeliveryDto);
-    VehicleDeliveryDto getVehicleDeliveryById(int id);
+    VehicleDeliveryDto getVehicleDeliveryById(Long id);
         List<VehicleDeliveryDto>  getAllvehicleDelivery();
-    VehicleDeliveryDto updateVehicleDelivery(int id,VehicleDeliveryDto vehicleDeliveryDto);
-        void deleteVehicleDelivery(int id);
+    VehicleDeliveryDto updateVehicleDelivery(Long id,VehicleDeliveryDto vehicleDeliveryDto);
+        void deleteVehicleDelivery(Long id);
 
 }
